+++ docs/ReleaseNotes.rst
@@ -31,6 +31,12 @@ infrastructure are described first, followed by tool-specific sections.
 Improvements to clang-tidy
 --------------------------
 
+- The 'nodecpp-naked-ptr-from-operator' check was removed
+
+- The 'nodecpp-naked-ptr-from-method' check was removed
+
+- The 'nodecpp-naked-ptr-from-function' check was removed
+
 - The 'nodecpp-naked-ptr-field' check was renamed to `nodecpp-var-decl
   <http://clang.llvm.org/extra/clang-tidy/checks/nodecpp-var-decl.html>`_